Added basic means of walking the filesystem
Created the first steps in the walk function, which will traverse a
filesystem and print all entries in a tree like form. Currently, the
current working directory is obtained and traversed. The number of files
and directories are tallied, and printed.

Also, made Counter struct public, along with its two methods.