Modified Walk Algorithm to Use a Single Iterator
The walk algorithm now creates a peekable iterator from the children of
each directory. This allows for executing a while loop which ensure the
list of children is not empty, while then peeking again to determine
which symbol in the tree is necessary. This should also be faster as the
elements are never aggregated in a single vector, but streamed fluidly
through iterators.