Consolidated Calls to `walk` Function
Defined a string within the scope of the while loop that initially
stores the prefix provided. Depending on whether there are additional
children elements following this one, a different character set may be
pushed onto the prefix. The `walk` function is then called from the
scope of the while loop, where the string may be accessed.