Consolidated Calls to `print_line` Function
Defined a string slice within the scope of the while loop that is
initially undefined. Depending on whether there are additional children
elements following this one, a different character set may be used as
the seperator between the tree prefix and the file name. The
`print_line` function is then called from the scope of the while loop,
where the string slice may be accessed.