Created the `all_files` Argument and Migrated App
Created the `all_files` argument to determine whether to display hidden
files or not. This argument is created via its own method, like the
others. Additionally, attaching the arguments to the app has been
delayed until after the app is constructed, in a seperate function. This
will allow for modular aggregation of arguments.