Modified `walk_dir` to Accept Function
Modified function `walk_dir` to accept a function that will act as a
predicate for filter. This will ensure that the iterator will deliver
proper results to the parsing system, i.e. they will not be filtered
after the `peek` call by the iterator. Also, the config may be parsed to
create a single predicate function, which may then be used for all calls
to `walk_dir`.